next update
added some features and fix buges

+++ Sergeev_DS_gun17_diplom/Assets/_Scripts/Interfaces/IDamageable.cs
@@ -1,11 +1,12 @@
 using System.Collections;
 using System.Collections.Generic;
+using Metroidvania.BaseUnit;
 using UnityEngine;
 namespace Metroidvania.Interfaces
 {
     public interface IDamageable
     {
-        void Damage(float amount);
+        void Damage(int amount);
     }
 }
